requirements.txt
pandas
requests
configparser
pycryptodomex
beautifulsoup4

